requirements.test.txt
pytest
pytest-cov==4.0.0
pytest-spec==3.2.0
pytest-asyncio==0.16.0
requests_mock
